Fix test failures
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/activemq/trunk@491857 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-jpa-store/src/main/java/org/apache/activemq/store/jpa/JPAMessageStore.java
@@ -92,17 +92,17 @@ public class JPAMessageStore implements MessageStore {
 	}
 
 	public int getMessageCount() throws IOException {
-		Integer rc;
+		Long rc;
 		EntityManager manager = adapter.beginEntityManager(null);
 		try {
 			Query query = manager.createQuery("select count(m) from StoredMessage m");
-			rc = (Integer) query.getSingleResult();
+			rc = (Long) query.getSingleResult();
 		} catch (Throwable e) {
 			adapter.rollbackEntityManager(null,manager);
 			throw IOExceptionSupport.create(e);
 		}
 		adapter.commitEntityManager(null,manager);
-		return rc;
+		return rc.intValue();
 	}
 
 	public String getMessageReference(MessageId identity) throws IOException {
